IPv4Subnet mask length is now an int, not an annoying byte

+++ src/main/java/com/github/maltalex/ineter/range/IPv4Range.java
@@ -181,7 +181,7 @@ public class IPv4Range implements IPRange<IPv4Address> {
 		}
 
 		int hostBits = Math.min(addrHostBits, hostBitsMax);
-		return IPv4Subnet.of(addr, (byte) (32 - hostBits));
+		return IPv4Subnet.of(addr, 32 - hostBits);
 	}
 
 	@Override
